Refactored parts; code cleanups

+++ src/main/java/com/defano/wyldcard/window/layouts/ShapesPalette.java
@@ -73,7 +73,7 @@ public class ShapesPalette extends WyldCardDialog<Object> implements Consumer<In
             thisShape.setEnabled(true);
         }
 
-        JButton selectedButton = getButtonForShape((int) newValue);
+        JButton selectedButton = getButtonForShape(newValue);
         if (selectedButton != null) {
             selectedButton.setEnabled(false);
         }
